% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_co2_vaqao.R
\name{decode_co2_vaqao}
\alias{decode_co2_vaqao}
\title{Decodificación de datos CO2 del sensor Watteco VAQAO}
\usage{
decode_co2_vaqao(dato_hex)
}
\arguments{
\item{dato_hex}{}
}
\value{
json
}
\description{
Decodifica los datos del sensor Watteco VAQAO. Como input, recibe la trama de datos en hexadecimal sin espacios.
}
\examples{
 decode_co2_vaqao(dato_hex)

}
